requirements.txt
flask==0.12
pg8000==1.10.6
requests==2.13.0
Fabric3==1.13.1.post1